Rename files; add test example
