+++ frontend/package.json
@@ -21,6 +21,7 @@
     "react-query": "^3.38.1"
   },
   "devDependencies": {
+    "@testing-library/dom": "^8.14.0",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/react-hooks": "^8.0.1",
@@ -39,6 +40,7 @@
     "eslint-plugin-react-hooks": "^4.5.0",
     "generate-react-cli": "^7.2.0",
     "jsdom": "^20.0.0",
+    "node-fetch": "^3.2.6",
     "prettier": "^2.6.2",
     "typescript": "^4.6.3",
     "vite": "2.9.12",
