+++ frontend/src/components/AsyncErrorBoundary/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AsyncErrorBoundary';
