+++ frontend/src/components/BananaSplit/BananaSplit.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+export interface BananaSplitProps {}
+
+const BananaSplit: FC<BananaSplitProps> = () => (
+  <div data-testid="BananaSplit">BananaSplit Component</div>
+);
+
+export default BananaSplit;
