+++ frontend/src/components/BananaSplit/index.ts
@@ -1,0 +1,2 @@
+export * from './BananaSplit';
+export { default } from './BananaSplit';
