+++ frontend/src/components/DelegationList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DelegationList';
