+++ frontend/src/components/DelegationList/models.ts
