+++ frontend/src/components/FeatureAuthAdmin/FeatureAuthAdmin.tsx
@@ -1,18 +1,18 @@
 import { useTranslation } from 'react-i18next';
 
-import AsyncErrorBoundary from '../async-error-boundary';
-import OrganisationDelegationRequests from '../organisation-delegation-requests';
+import AsyncErrorBoundary from '../AsyncErrorBoundary';
+import OrganisationDelegationRequests from '../OrganisationDelegationRequests';
 
 const App = () => {
   const { t } = useTranslation('common');
 
   return (
-    <>
+    <div data-testid="FeatureAuthAdmin">
       <h1>{t('greeting')}…</h1>
       <AsyncErrorBoundary>
         <OrganisationDelegationRequests />
       </AsyncErrorBoundary>
-    </>
+    </div>
   );
 };
