+++ frontend/src/components/OrganisationDelegationRequests/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OrganisationDelegationRequests';
