+++ frontend/src/components/SuspendLoadLocale/SuspendLoadLocale.tsx
