--- frontend/src/components/async-error-boundary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './async-error-boundary';
