--- frontend/src/components/delegation-list/index.ts
@@ -1,1 +1,0 @@
-export { default } from './delegation-list';
