--- frontend/src/components/feature-auth-admin/index.ts
@@ -1,1 +1,0 @@
-export { default } from './feature-auth-admin';
