--- frontend/src/components/organisation-delegation-requests/index.ts
@@ -1,1 +1,0 @@
-export { default } from './organisation-delegation-requests';
