--- frontend/src/components/suspend-load-locale/index.ts
@@ -1,1 +1,0 @@
-export { default } from './suspend-load-locale';
